firebaseImage this is the image to firebase image
